gift_certificate.sql - price type changed to decimal

+++ repository/src/test/resources/sql_scripts/gift_certificate.sql
@@ -3,7 +3,7 @@ create table gift_certificate
     id                      BIGINT AUTO_INCREMENT,
     name                    varchar(100)  NOT NULL,
     description             VARCHAR(200) NOT NULL,
-    price                   DOUBLE NOT NULL,
+    price                   DECIMAL NOT NULL,
     duration                INT NOT NULL,
     create_date             TIMESTAMP NOT NULL,
     last_update_date        TIMESTAMP NOT NULL,
